v0.1.3
- bugs fixes
- Pause

+++ projAbmooction/Assets/Scripts/Movement.cs
@@ -27,7 +27,7 @@ class Movement
         TargetPos = finalPos;
     }
 
-    public void ChangeSpeed(int speed)
+    public void ChangeSpeed(float speed)
     {
         Speed = speed;
     }
